pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "annpack"
version = "0.1.5"
description = "ANNPack builder and tools"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
dependencies = [
  "polars",
  "faiss-cpu",
  "numpy",
  "cryptography",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/Arjun2729/ANNPACK"
Issues = "https://github.com/Arjun2729/ANNPACK/issues"

[project.optional-dependencies]
dev = ["pytest", "ruff", "mkdocs", "hypothesis", "build", "twine", "mypy", "pre-commit"]
embed = ["sentence-transformers", "torch", "datasets"]
registry = ["fastapi", "uvicorn", "pyjwt", "python-multipart", "slowapi", "limits"]

[project.scripts]
annpack = "annpack.cli:main"
annpack-build = "annpack.cli:main"

[tool.setuptools]
package-dir = {"" = "python"}
packages = ["annpack", "annpack.ui"]
package-data = {"annpack" = ["ui/*.html", "py.typed"]}

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
files = [
  "python/annpack/api.py",
  "python/annpack/cli.py",
  "python/annpack/packset.py",
  "python/annpack/reader.py",
  "python/annpack/verify.py",
]
ignore_missing_imports = true
strict = true
warn_unused_ignores = true
